requirements.txt
torchaudio
tensorboard
einops
matplotlib
pyyaml
tqdm
torchcodec
soundfile
geoopt
scikit-learn
torchvision
hypll
geoopt
torchmetrics
torch-fidelity
